backend/requirements.txt
beautifulsoup4==4.8.0
certifi==2019.11.28
chardet==3.0.4
Django==2.2.9
djangorestframework==3.11.0
idna==2.8
pytz==2019.3
requests==2.22.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.7
